third_party/blink/web_tests/fast/css/link-alternate-stylesheet-4.html

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.1//EN" "http://www.w3.org/TR/xhtml11/DTD/xhtml11.dtd">
<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en"><head>
  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
  <link rel="stylesheet" href="resources/basic.css" type="text/css">
  <link rel="stylesheet" href="resources/small.css" type="text/css">
  <link rel="stylesheet" href="resources/normal.css" type="text/css">
  <link rel="stylesheet" href="resources/high.css" title="Layout" type="text/css">
  <script src="../../resources/js-test.js"></script>
</head>
<body>
  <h1 id="firstH1">Alternate stylesheets</h1>
  
  <p id="textLine">This document has 4 stylesheets, with no alternate stylesheets but the "high.css" stylesheet now
    has a title attribute.</p>
  
  <ul>
    <li>Have a red border at the top of the page</li>
    <li>12px text</li>
  </ul>

  <script>
    description("Alternate stylesheet (link) title test 4");

    if (window.testRunner)
      testRunner.dumpAsText();

    window.onload = function () {
      element = document.getElementById("firstH1");
      shouldBe("getComputedStyle(element).getPropertyValue('text-decoration')", "'none'");

      element = document.getElementById("textLine");
      shouldBe("getComputedStyle(element).getPropertyValue('font-size')", "'12px'");

      shouldBe("getComputedStyle(document.body).getPropertyValue('border-top-color')", "'rgb(139, 0, 0)'");

      wasPostTestScriptParsed = true;
      finishJSTest();
    }
  </script>

</body>
</html>
