third_party/WebKit/LayoutTests/fast/css/link-alternate-stylesheet-4-expected.txt

Alternate stylesheet (link) title test 4

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getComputedStyle(element).getPropertyValue('text-decoration') is 'none solid rgb(51, 51, 51)'
PASS getComputedStyle(element).getPropertyValue('font-size') is '12px'
PASS getComputedStyle(document.body).getPropertyValue('border-top-color') is 'rgb(139, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE

Alternate stylesheets

This document has 4 stylesheets, with no alternate stylesheets but the "high.css" stylesheet now has a title attribute.

Have a red border at the top of the page
12px text
